Remove mention of formal and actual parameters.
This section appeared confusing, especially to devs seeing this for the first
time. I've gone in and removed references to formal params and actual params as
the main point of this should be arguments = things you use when you call the
function and parameters = things you use when you declare the function.